c99: c0/c1 node-union, sh port, adb/ar/cpp/libc/headers modernization
Commit the c99-tier modernization accumulated across the porting passes:

- c0/c1 full-union node refactor (c0.h/c00-c04.c/c1.h/c10-c12.c).
- sh port (args/cmd/ctype/expand/fault/main/msg/name/service/word/xec.c + the
  10 defs/mac/mode/name/stak/sym/... headers).
- make/misc.c and rm.c host-adaptation fixes.
- adb source (20 files) + ar.c + cpp cpy.c/yylex.c.
- c99/usr/include/: the modernized V7 headers (a.out.h .. sys/*.h).
- libc/crt/udiv.s + urem.s (the pcc unsigned-division helpers).

+++ c99/usr/src/cmd/sh/fault.c
@@ -54,7 +54,7 @@ int16_t ignsig(int16_t n)
 {
 	REG INT		s, i;
 
-	IF (s=signal(i=n,1)&01)==0
+	IF (s=(int16_t)signal(i=n,1)&01)==0
 	THEN	trapflg[i] |= SIGMOD;
 	FI
 	return(s);
